Create add_worldfile.rb
At work today I had to learn how to copy a file and rename it based on certain files within the directory in Python.  This is my translation of that simple program to Ruby.  

This is my first attempt at creating a Ruby program to do what I want it to do, instead of simply following some code along, or doing the prescribed homework that goes with those code alongs.